Use DisplayName instead of methodName
This change uses the TestResult.DisplayName or TestResult.TestCase.DisplayName as the value of the name attribute of testCase elements.

This has the following benefits over using the name of the test method:
1. It includes the namespace and type name for disambiguation of similarly named tests
2. In case of theory tests, it includes the values of the parameters for disambiguation

The change should match the behavior of the TrxLogger, see https://github.com/Microsoft/vstest/blob/master/src/Microsoft.TestPlatform.Extensions.TrxLogger/TrxLogger.cs#L546 for details.